+++ analyzer/GlobalUsings.cs
@@ -1,6 +1,7 @@
 global using System;
-global using System.Linq;
+global using System.Collections.Frozen;
 global using System.Collections.Generic;
 global using System.Collections.Immutable;
-global using System.Collections.Frozen;
+global using System.Linq;
 global using System.Threading.Tasks;
+global using static Ametrin.Optional.Analyzer.Helper;