+++ analyzer/Helper.cs
@@ -10,4 +10,6 @@ internal static class Helper
     internal static bool IsOptionalType(ITypeSymbol type) => type is { Name: "Option" or "Result" or "ErrorState" or "RefOption", ContainingAssembly.Name: "Ametrin.Optional" };
     internal static bool IsResultType(ITypeSymbol type) => type is { Name: "Result", ContainingAssembly.Name: "Ametrin.Optional" };
     internal static bool IsDefaultLiteral(IOperation operation) => operation.Syntax.IsKind(SyntaxKind.DefaultLiteralExpression);
+    internal static bool HasAttribute(ISymbol symbol, Func<INamedTypeSymbol?, bool> predicate) => symbol.GetAttributes().Any(data => predicate(data.AttributeClass));
+    internal static bool IsGenerateParsingAttribute(INamedTypeSymbol? attribute) => attribute is { Name: "GenerateParsingAttribute", ContainingAssembly.Name: "Ametrin.Optional" };
 }